Рабочая версия проекта без дизайна

+++ Matem/Matem/StudentTasks1.cs
@@ -28,42 +28,46 @@ namespace Matem
 
         private void Theme1_Click(object sender, EventArgs e)
         {
-            Theme1_Zadachi form = new Theme1_Zadachi();
-            List<Mission> any = new List<Mission>();
-            XmlSerializer diser = new XmlSerializer(typeof(List<Mission>));
-            using (FileStream fs = new FileStream("bank.xml", FileMode.OpenOrCreate))
+            if(Theme1.Text!="---")
             {
-                any = (List<Mission>)diser.Deserialize(fs);
-            }
-            Nazvanie_Theme nazvanie = new Nazvanie_Theme();
-            XmlSerializer formater = new XmlSerializer(typeof(Nazvanie_Theme));
-            using (FileStream fs = new FileStream("theme.xml", FileMode.OpenOrCreate))
-            {
-                nazvanie = (Nazvanie_Theme)formater.Deserialize(fs);
-            }
-            List<Mission> list = new List<Mission>();
-            foreach(var item in any)
-            {
-                if (item.Theme == nazvanie.Name)
+                Theme1_Zadachi form = new Theme1_Zadachi();
+                List<Mission> any = new List<Mission>();
+                XmlSerializer diser = new XmlSerializer(typeof(List<Mission>));
+                using (FileStream fs = new FileStream("bank.xml", FileMode.OpenOrCreate))
                 {
-                    list.Add(item);
-                    
+                    any = (List<Mission>)diser.Deserialize(fs);
                 }
-            }
-            XmlSerializer formatterlist = new XmlSerializer(typeof(List<Mission>));
+                /*Nazvanie_Theme nazvanie = new Nazvanie_Theme();
+                XmlSerializer formater = new XmlSerializer(typeof(Nazvanie_Theme));
+                using (FileStream fs = new FileStream("theme.xml", FileMode.OpenOrCreate))
+                {
+                    nazvanie = (Nazvanie_Theme)formater.Deserialize(fs);
+                }*/
+                List<Mission> list = new List<Mission>();
+                foreach (var item in any)
+                {
+                    if (item.Theme == Theme1.Text)
+                    {
+                        list.Add(item);
+                    }
+                }
+                XmlSerializer formatterlist = new XmlSerializer(typeof(List<Mission>));
 
 
-            if(File.Exists("listmission.xml"))
-            {
-                File.Delete("listmission.xml");
+                if (File.Exists("listmission.xml"))
+                {
+                    File.Delete("listmission.xml");
+                }
+                using (FileStream fs = new FileStream("listmission.xml", FileMode.OpenOrCreate))
+                {
+                    formatterlist.Serialize(fs, list);
+                }
+                form.Show();
+                this.Hide();
             }
-            using (FileStream fs = new FileStream("listmission.xml", FileMode.OpenOrCreate))
-            {
-                formatterlist.Serialize(fs, list);
-            }
-            form.Show();
-            this.Hide();
+            
         }
+
         void SearchThemes(List<Mission> any, List<string> thm) //метод находит все уникальные темы в хранилище
         {
            
@@ -73,6 +77,375 @@ namespace Matem
                 {
                     thm.Add(item.Theme);
                 }
+            }
+        }
+
+        private void Theme2_Click(object sender, EventArgs e)
+        {
+            if (Theme2.Text != "---")
+            {
+                Theme1_Zadachi form = new Theme1_Zadachi();
+                List<Mission> any = new List<Mission>();
+                XmlSerializer diser = new XmlSerializer(typeof(List<Mission>));
+                using (FileStream fs = new FileStream("bank.xml", FileMode.OpenOrCreate))
+                {
+                    any = (List<Mission>)diser.Deserialize(fs);
+                }
+                /*Nazvanie_Theme nazvanie = new Nazvanie_Theme();
+                XmlSerializer formater = new XmlSerializer(typeof(Nazvanie_Theme));
+                using (FileStream fs = new FileStream("theme.xml", FileMode.OpenOrCreate))
+                {
+                    nazvanie = (Nazvanie_Theme)formater.Deserialize(fs);
+                }*/
+                List<Mission> list = new List<Mission>();
+                foreach (var item in any)
+                {
+                    if (item.Theme == Theme2.Text)
+                    {
+                        list.Add(item);
+                    }
+                }
+                XmlSerializer formatterlist = new XmlSerializer(typeof(List<Mission>));
+
+
+                if (File.Exists("listmission.xml"))
+                {
+                    File.Delete("listmission.xml");
+                }
+                using (FileStream fs = new FileStream("listmission.xml", FileMode.OpenOrCreate))
+                {
+                    formatterlist.Serialize(fs, list);
+                }
+                form.Show();
+                this.Hide();
+            }
+        }
+
+        private void Theme3_Click(object sender, EventArgs e)
+        {
+            if (Theme3.Text != "---")
+            {
+                Theme1_Zadachi form = new Theme1_Zadachi();
+                List<Mission> any = new List<Mission>();
+                XmlSerializer diser = new XmlSerializer(typeof(List<Mission>));
+                using (FileStream fs = new FileStream("bank.xml", FileMode.OpenOrCreate))
+                {
+                    any = (List<Mission>)diser.Deserialize(fs);
+                }
+                /*Nazvanie_Theme nazvanie = new Nazvanie_Theme();
+                XmlSerializer formater = new XmlSerializer(typeof(Nazvanie_Theme));
+                using (FileStream fs = new FileStream("theme.xml", FileMode.OpenOrCreate))
+                {
+                    nazvanie = (Nazvanie_Theme)formater.Deserialize(fs);
+                }*/
+                List<Mission> list = new List<Mission>();
+                foreach (var item in any)
+                {
+                    if (item.Theme == Theme3.Text)
+                    {
+                        list.Add(item);
+                    }
+                }
+                XmlSerializer formatterlist = new XmlSerializer(typeof(List<Mission>));
+
+
+                if (File.Exists("listmission.xml"))
+                {
+                    File.Delete("listmission.xml");
+                }
+                using (FileStream fs = new FileStream("listmission.xml", FileMode.OpenOrCreate))
+                {
+                    formatterlist.Serialize(fs, list);
+                }
+                form.Show();
+                this.Hide();
+            }
+        }
+
+        private void Theme4_Click(object sender, EventArgs e)
+        {
+            if (Theme4.Text != "---")
+            {
+                Theme1_Zadachi form = new Theme1_Zadachi();
+                List<Mission> any = new List<Mission>();
+                XmlSerializer diser = new XmlSerializer(typeof(List<Mission>));
+                using (FileStream fs = new FileStream("bank.xml", FileMode.OpenOrCreate))
+                {
+                    any = (List<Mission>)diser.Deserialize(fs);
+                }
+                /*Nazvanie_Theme nazvanie = new Nazvanie_Theme();
+                XmlSerializer formater = new XmlSerializer(typeof(Nazvanie_Theme));
+                using (FileStream fs = new FileStream("theme.xml", FileMode.OpenOrCreate))
+                {
+                    nazvanie = (Nazvanie_Theme)formater.Deserialize(fs);
+                }*/
+                List<Mission> list = new List<Mission>();
+                foreach (var item in any)
+                {
+                    if (item.Theme == Theme4.Text)
+                    {
+                        list.Add(item);
+                    }
+                }
+                XmlSerializer formatterlist = new XmlSerializer(typeof(List<Mission>));
+
+
+                if (File.Exists("listmission.xml"))
+                {
+                    File.Delete("listmission.xml");
+                }
+                using (FileStream fs = new FileStream("listmission.xml", FileMode.OpenOrCreate))
+                {
+                    formatterlist.Serialize(fs, list);
+                }
+                form.Show();
+                this.Hide();
+            }
+        }
+
+        private void Theme5_Click(object sender, EventArgs e)
+        {
+            if (Theme5.Text != "---")
+            {
+                Theme1_Zadachi form = new Theme1_Zadachi();
+                List<Mission> any = new List<Mission>();
+                XmlSerializer diser = new XmlSerializer(typeof(List<Mission>));
+                using (FileStream fs = new FileStream("bank.xml", FileMode.OpenOrCreate))
+                {
+                    any = (List<Mission>)diser.Deserialize(fs);
+                }
+                /*Nazvanie_Theme nazvanie = new Nazvanie_Theme();
+                XmlSerializer formater = new XmlSerializer(typeof(Nazvanie_Theme));
+                using (FileStream fs = new FileStream("theme.xml", FileMode.OpenOrCreate))
+                {
+                    nazvanie = (Nazvanie_Theme)formater.Deserialize(fs);
+                }*/
+                List<Mission> list = new List<Mission>();
+                foreach (var item in any)
+                {
+                    if (item.Theme == Theme5.Text)
+                    {
+                        list.Add(item);
+                    }
+                }
+                XmlSerializer formatterlist = new XmlSerializer(typeof(List<Mission>));
+
+
+                if (File.Exists("listmission.xml"))
+                {
+                    File.Delete("listmission.xml");
+                }
+                using (FileStream fs = new FileStream("listmission.xml", FileMode.OpenOrCreate))
+                {
+                    formatterlist.Serialize(fs, list);
+                }
+                form.Show();
+                this.Hide();
+            }
+        }
+
+        private void Theme6_Click(object sender, EventArgs e)
+        {
+            if (Theme6.Text != "---")
+            {
+                Theme1_Zadachi form = new Theme1_Zadachi();
+                List<Mission> any = new List<Mission>();
+                XmlSerializer diser = new XmlSerializer(typeof(List<Mission>));
+                using (FileStream fs = new FileStream("bank.xml", FileMode.OpenOrCreate))
+                {
+                    any = (List<Mission>)diser.Deserialize(fs);
+                }
+                /*Nazvanie_Theme nazvanie = new Nazvanie_Theme();
+                XmlSerializer formater = new XmlSerializer(typeof(Nazvanie_Theme));
+                using (FileStream fs = new FileStream("theme.xml", FileMode.OpenOrCreate))
+                {
+                    nazvanie = (Nazvanie_Theme)formater.Deserialize(fs);
+                }*/
+                List<Mission> list = new List<Mission>();
+                foreach (var item in any)
+                {
+                    if (item.Theme == Theme6.Text)
+                    {
+                        list.Add(item);
+                    }
+                }
+                XmlSerializer formatterlist = new XmlSerializer(typeof(List<Mission>));
+
+
+                if (File.Exists("listmission.xml"))
+                {
+                    File.Delete("listmission.xml");
+                }
+                using (FileStream fs = new FileStream("listmission.xml", FileMode.OpenOrCreate))
+                {
+                    formatterlist.Serialize(fs, list);
+                }
+                form.Show();
+                this.Hide();
+            }
+        }
+
+        private void Theme7_Click(object sender, EventArgs e)
+        {
+            if (Theme7.Text != "---")
+            {
+                Theme1_Zadachi form = new Theme1_Zadachi();
+                List<Mission> any = new List<Mission>();
+                XmlSerializer diser = new XmlSerializer(typeof(List<Mission>));
+                using (FileStream fs = new FileStream("bank.xml", FileMode.OpenOrCreate))
+                {
+                    any = (List<Mission>)diser.Deserialize(fs);
+                }
+                /*Nazvanie_Theme nazvanie = new Nazvanie_Theme();
+                XmlSerializer formater = new XmlSerializer(typeof(Nazvanie_Theme));
+                using (FileStream fs = new FileStream("theme.xml", FileMode.OpenOrCreate))
+                {
+                    nazvanie = (Nazvanie_Theme)formater.Deserialize(fs);
+                }*/
+                List<Mission> list = new List<Mission>();
+                foreach (var item in any)
+                {
+                    if (item.Theme == Theme7.Text)
+                    {
+                        list.Add(item);
+                    }
+                }
+                XmlSerializer formatterlist = new XmlSerializer(typeof(List<Mission>));
+
+
+                if (File.Exists("listmission.xml"))
+                {
+                    File.Delete("listmission.xml");
+                }
+                using (FileStream fs = new FileStream("listmission.xml", FileMode.OpenOrCreate))
+                {
+                    formatterlist.Serialize(fs, list);
+                }
+                form.Show();
+                this.Hide();
+            }
+        }
+
+        private void Theme8_Click(object sender, EventArgs e)
+        {
+            if (Theme8.Text != "---")
+            {
+                Theme1_Zadachi form = new Theme1_Zadachi();
+                List<Mission> any = new List<Mission>();
+                XmlSerializer diser = new XmlSerializer(typeof(List<Mission>));
+                using (FileStream fs = new FileStream("bank.xml", FileMode.OpenOrCreate))
+                {
+                    any = (List<Mission>)diser.Deserialize(fs);
+                }
+                /*Nazvanie_Theme nazvanie = new Nazvanie_Theme();
+                XmlSerializer formater = new XmlSerializer(typeof(Nazvanie_Theme));
+                using (FileStream fs = new FileStream("theme.xml", FileMode.OpenOrCreate))
+                {
+                    nazvanie = (Nazvanie_Theme)formater.Deserialize(fs);
+                }*/
+                List<Mission> list = new List<Mission>();
+                foreach (var item in any)
+                {
+                    if (item.Theme == Theme8.Text)
+                    {
+                        list.Add(item);
+                    }
+                }
+                XmlSerializer formatterlist = new XmlSerializer(typeof(List<Mission>));
+
+
+                if (File.Exists("listmission.xml"))
+                {
+                    File.Delete("listmission.xml");
+                }
+                using (FileStream fs = new FileStream("listmission.xml", FileMode.OpenOrCreate))
+                {
+                    formatterlist.Serialize(fs, list);
+                }
+                form.Show();
+                this.Hide();
+            }
+        }
+
+        private void Theme9_Click(object sender, EventArgs e)
+        {
+            if (Theme9.Text != "---")
+            {
+                Theme1_Zadachi form = new Theme1_Zadachi();
+                List<Mission> any = new List<Mission>();
+                XmlSerializer diser = new XmlSerializer(typeof(List<Mission>));
+                using (FileStream fs = new FileStream("bank.xml", FileMode.OpenOrCreate))
+                {
+                    any = (List<Mission>)diser.Deserialize(fs);
+                }
+                /*Nazvanie_Theme nazvanie = new Nazvanie_Theme();
+                XmlSerializer formater = new XmlSerializer(typeof(Nazvanie_Theme));
+                using (FileStream fs = new FileStream("theme.xml", FileMode.OpenOrCreate))
+                {
+                    nazvanie = (Nazvanie_Theme)formater.Deserialize(fs);
+                }*/
+                List<Mission> list = new List<Mission>();
+                foreach (var item in any)
+                {
+                    if (item.Theme == Theme9.Text)
+                    {
+                        list.Add(item);
+                    }
+                }
+                XmlSerializer formatterlist = new XmlSerializer(typeof(List<Mission>));
+
+
+                if (File.Exists("listmission.xml"))
+                {
+                    File.Delete("listmission.xml");
+                }
+                using (FileStream fs = new FileStream("listmission.xml", FileMode.OpenOrCreate))
+                {
+                    formatterlist.Serialize(fs, list);
+                }
+                form.Show();
+                this.Hide();
+            }
+        }
+
+        private void Theme10_Click(object sender, EventArgs e)
+        {
+            if (Theme10.Text != "---")
+            {
+                Theme1_Zadachi form = new Theme1_Zadachi();
+                List<Mission> any = new List<Mission>();
+                XmlSerializer diser = new XmlSerializer(typeof(List<Mission>));
+                using (FileStream fs = new FileStream("bank.xml", FileMode.OpenOrCreate))
+                {
+                    any = (List<Mission>)diser.Deserialize(fs);
+                }
+                /*Nazvanie_Theme nazvanie = new Nazvanie_Theme();
+                XmlSerializer formater = new XmlSerializer(typeof(Nazvanie_Theme));
+                using (FileStream fs = new FileStream("theme.xml", FileMode.OpenOrCreate))
+                {
+                    nazvanie = (Nazvanie_Theme)formater.Deserialize(fs);
+                }*/
+                List<Mission> list = new List<Mission>();
+                foreach (var item in any)
+                {
+                    if (item.Theme == Theme10.Text)
+                    {
+                        list.Add(item);
+                    }
+                }
+                XmlSerializer formatterlist = new XmlSerializer(typeof(List<Mission>));
+
+
+                if (File.Exists("listmission.xml"))
+                {
+                    File.Delete("listmission.xml");
+                }
+                using (FileStream fs = new FileStream("listmission.xml", FileMode.OpenOrCreate))
+                {
+                    formatterlist.Serialize(fs, list);
+                }
+                form.Show();
+                this.Hide();
             }
         }
     }
